Add files via upload
Cool ... lock is neat but The more locks you are acquire at one time, the more you lose the advantages of concurrency.
Locks don’t lock anything. They are just flags and can be ignored. It is a cooperative tool, not an enforced tool.
 locks should be considered a low level primitive that is difficult to reason about in non-trivial examples.
The general rule is “thread after you fork not before”. Otherwise, the locks used by the thread executor will get duplicated across processes. If one of those processes dies while it has the lock, all of the other processes using that lock will deadlock.
 Previous